requirements.txt
pynacl
requests
numpy
Pillow
